Prepare new code entry-points
Code is moved into more idiomatic source directories. Functionality
will need to be reenabled.
